Added functionality to add users to a newly created group

+++ schema/000001_init.up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE users
 (
     user_id      SERIAL PRIMARY KEY            NOT NULL,
-    chat_id      INT                           NOT NULL,
+    chat_id      BIGINT                         NOT NULL,
     user_name    VARCHAR(50) DEFAULT 'no_name' NOT NULL,
     language     VARCHAR(5)  DEFAULT 'EN'      NOT NULL,
     status       INT                           NOT NULL DEFAULT 0 NOT NULL,
